Abstracted out Rating component.
