+++ Interactive-rating-component/src/index.css
@@ -49,48 +49,6 @@ body {
   gap: 1rem;
 }
 
-.interactive-rating-display {
-  display: flex;
-  flex-direction: column;
-  width: 375px;
-  background-color: var(--clr-veryDarkBlue);
-  padding: 2rem;
-  border-radius: 1rem;
-  gap: 1.5rem;
-}
-
-.star-image {
-  align-self: flex-start;
-  padding: 15px;
-  border-radius: 50%;
-  background: var(--clr-darkBlue);
-  cursor: pointer;
-}
-
-.rating-message {
-  display: flex;
-  flex-direction: column;
-  gap: 0.25rem;
-}
-
-.display-title {
-  color: var(--clr-white);
-  font-weight: var(--fw-bold);
-}
-
-.display-message {
-  color: var(--clr-lightGrey);
-  font-weight: var(--fw-regular);
-  font-size: var(--font-paragraph);
-}
-
-.rating-buttons {
-  display: flex;
-  flex-direction: row;
-  gap: 1rem;
-  justify-content: space-between;
-}
-
 .attribution {
   color: var(--clr-white);
 }
